refactor: 使い回すスライスのリセットを internal/sliceutil に切り出す
iterator.resetByteSlices と output.resetStringSlice が同じロジックの重複に
なっていたので、ジェネリックな sliceutil.Reset[T] にまとめる。

パッケージ名を slices ではなく sliceutil にしたのは、将来 stdlib の slices を
import したときに衝突させないため。
